Manage Listings Page & User Authorization

+++ app/Http/Controllers/ListingController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Validation\Rule;
@@ -49,10 +50,7 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
-
         $formFields['user_id'] = auth()->id();
-        // dd($formFields);
-
 
         Listing::create($formFields);
         return redirect('/')->with('message', 'Listing create successfuly!');
@@ -68,6 +66,9 @@ class ListingController extends Controller
     //Update
     public function update(Request $request, Listing $listing)
     {
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action!');
+        }
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -82,14 +83,23 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo')->store('logos', 'public');
         }
         $listing->update($formFields);
-        return back()->with('message', 'Listing update successfuly!');
+        return redirect('/')->with('message', 'Listing update successfuly!');
     }
 
     // Delete Listing
 
     public function destroy(Listing $listing)
     {
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized Action!');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Lisitng deletrd successfully');
+    }
+
+    public function manage()
+    {
+        return view('listings.manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
